balancing game mechanics
changed worth of some critters
changed damage/health/speed of some critters
changed firing speed damage of some towers

+++ src/towers/Tower.java
@@ -21,7 +21,7 @@ public abstract class Tower {
 	private double yPos;
 	protected boolean freezeTower = false;
 	private Critter targetCritter;
-	protected int reloadTime;
+	protected double reloadTime;
 	private long lastAttackTime;
 	private double angleOfRotation;
 	protected type towerType;
